hyperliquid scalping agent/requirements.txt
hyperliquid-python-sdk>=0.22.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
pandas-ta>=0.3.14b
aiohttp>=3.9.0
requests>=2.31.0
